--- src/main/java/StringAddCalculator.java
@@ -1,11 +1,0 @@
-import vo.*;
-
-public class StringAddCalculator {
-
-    public static int splitAndSum(String value) {
-        SplitTarget target = SplitTarget.of(value);
-        SplitStrings splitStrings =  target.split();
-        PositiveIntegers positiveIntegers = splitStrings.toIntegers();
-        return positiveIntegers.sum();
-    }
-}
