+++ src/main/java/com/nextstep/camp/racing/RacingCarApplication.java
@@ -1,0 +1,10 @@
+package com.nextstep.camp.racing;
+
+import com.nextstep.camp.racing.config.ApplicationInitializer;
+
+public class RacingCarApplication {
+
+    public static void main(String[] args) {
+        ApplicationInitializer.initialize();
+    }
+}
