+++ src/main/java/com/nextstep/camp/racing/common/utils/StringUtils.java
@@ -1,0 +1,18 @@
+package com.nextstep.camp.racing.common.utils;
+
+import com.nextstep.camp.racing.domain.exception.UtilInstantiationException;
+
+public class StringUtils {
+
+    private StringUtils() {
+        throw new UtilInstantiationException();
+    }
+
+    public static boolean hasText(String value) {
+        return value == null || value.isBlank();
+    }
+
+    public static boolean hasNotText(String value) {
+        return !hasText(value);
+    }
+}
