+++ src/main/java/com/nextstep/camp/racing/infrastructure/view/Element.java
@@ -1,0 +1,9 @@
+package com.nextstep.camp.racing.infrastructure.view;
+
+public interface Element {
+    void action();
+    String getName();
+    String getLabel();
+    String getValue();
+}
+
