+++ src/main/java/com/nextstep/camp/racing/infrastructure/view/View.java
@@ -1,0 +1,5 @@
+package com.nextstep.camp.racing.infrastructure.view;
+
+public interface View {
+    ViewData toViewData();
+}
