+++ src/main/java/com/nextstep/camp/racing/infrastructure/view/ViewData.java
@@ -1,0 +1,19 @@
+package com.nextstep.camp.racing.infrastructure.view;
+
+import java.util.Map;
+
+public class ViewData {
+    private final Map<String, String> data;
+
+    private ViewData(Map<String, String> data) {
+        this.data = data;
+    }
+
+    public static ViewData of(Map<String, String> data) {
+        return new ViewData(data);
+    }
+
+    public String get(String key) {
+        return data.get(key);
+    }
+}
