+++ src/main/java/com/nextstep/camp/racing/presentation/RacingController.java
@@ -1,0 +1,4 @@
+package com.nextstep.camp.racing.presentation;
+
+public class RacingController {
+}
