--- src/main/java/exception/NegativeNumberException.java
@@ -1,10 +1,0 @@
-package exception;
-
-public class NegativeNumberException extends RuntimeException {
-
-    private static final String MESSAGE = "음수는 할당할 수 없습니다.";
-
-    public NegativeNumberException() {
-        super(MESSAGE);
-    }
-}
