--- src/main/java/exception/NotNumberException.java
@@ -1,10 +1,0 @@
-package exception;
-
-public class NotNumberException extends RuntimeException {
-
-    private static final String MESSAGE = "숫자가 아닌 값은 할당할 수 없습니다.";
-
-    public NotNumberException() {
-        super(MESSAGE);
-    }
-}
