--- src/main/java/exception/ViewDataNotFoundException.java
@@ -1,9 +1,0 @@
-package exception;
-
-public class ViewDataNotFoundException extends RuntimeException {
-    private static final String MESSAGE = "ViewData가 없습니다.";
-
-    public ViewDataNotFoundException() {
-        super(MESSAGE);
-    }
-}
