--- src/main/java/view/Element.java
@@ -1,11 +1,0 @@
-package view;
-
-import java.util.Map;
-
-public interface Element {
-    void action();
-    String getName();
-    String getLabel();
-    String getValue();
-}
-
