--- src/main/java/view/View.java
@@ -1,7 +1,0 @@
-package view;
-
-import java.util.Map;
-
-public interface View {
-    ViewData toViewData();
-}
